Force os-client-config to use ipv4
After we removed ipv6-preferred support it nodepool, we started to
default to ipv6 for testing. However, because the current state of
ipv6 in nova / neutron / glean, ipv6 is not working properly for
glean.

Revert to ipv4 testing only, until we are ready to dive down the
rabbit hole of ipv6.

Change-Id: I7dece93671fbff5a1260fe1721f2d1b4bf337e72
Signed-off-by: Paul Belanger <pabelanger@redhat.com>