Wait for lock when performing manual hold
So that an operator can choose to hold a node while it is in use,
update the 'nodepool hold' command to wait for the lock if that
is required.  The locking algorithm is ordered, so as long as this
command is run while the node is locked while in use, this command
will be able to obtain the lock on the node before the node
deleter does.

This means it will sit in the foreground until the node is locked,
which, in some cases, may be hours.  Perhaps that's not ideal,
but processes are easy to background in those cases, or abort if
the operator changes their mind.  It does provide an avenue for
notification that the node has been unlocked by returning to the
shell when complete.

Change-Id: I89721cbe1e53b30bb0e2b5f5440870df958cf2c1