Update patch set 1
Patch Set 1:

(2 comments)

I'm on the fence on this.  As Paul points out, this isn't actually used at all by nodepool.  It's just passed on to further users (e.g., zuul).  However, nodepool passes along other information it doesn't use either (ip addresses, ssh keys), and as it is responsible for generating the images, it seems natural to add this here.  I think the discomfort comes because this is the only item of information that we would be both "passing through" nodepool but originitating in its config file.

There is an accomodation in the zuulv3 spec for configuring zuul to use hosts which are statically defined in zuul (principally for CD -- so you can have zuul run a job on a production server without adding that to nodepool).  We haven't implemented it yet, but if/when we do, presumably we will need to configure the username as well.  We could extend that to general node configuration in zuul.  That's likely to be more repetitive, awkward, and error prone than simply using the value from nodepool.  And we can always do both, with the value from nodepool being a default.

All of that to say that I think I'm leaning slightly toward agreeing that this is the way we should go.

(A side node: we probably could have just added this to the config files where we actually verified the field in tests.  I think we have a bit too much needless boilerplate repitition in the config files and I'd like to reduce it.)

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0