Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

The thing called "unmanagedimage" is referenced by the thing called "image-name", so I think it would be a good idea to call them the same thing (in the way that we use "diskimage" both places).

I think it would be reasonable to call them both "image" (so the section in pool is "images" and the attribute on pool-label is "image").  But Monty's suggestion of "cloud-images" / "cloud-image" seems less ambiguous (I modified it to add a hyphen -- the only reason diskimage is not hyphenated is because it's the producte of the "diskimage-builder" program, but I don't think "cloudimage" is a word).

Other than that, and the test fixture suggestion, this looks good, thanks!

Patch-set: 1
Label: Code-Review=-1