Support homing diskimage builds to specific builder
In the case of openstack-infra, we'd like to bring a new
nodepool-builder online for xenial. This is because opensuse-minimal
build dependencies only support xenial at the moment. However, we do
not want to build opensuse-minimal on our existing trust
nodepool-buidlers.

This patch solves this by creating a builders field for diskimages,
allowing an operator to select which hosts builds will run on. If
empty, no filtering is enable and all nodepool-builders will attempt
to build the diskimage.

Change-Id: I066b22b4afd601c8f0020ad0ab306e625b73e231
Signed-off-by: Paul Belanger <pabelanger@redhat.com>