Copy nodepool logs to devstack log dir
This creates a new post test hook under devstack/gate.  The original
post test hook (which runs our nodepool sanity checks) is moved into
that directory as well.  The new post test hook calls that, and then,
regardless of whether that succeeded, copies the nodepool log files
into the devstack-gate logs directory.

We can change project-config to point to the new post test hook script
and then delete the original.

It also updates the nodepool config to write two logging config files,
one for the builder and one for the launcher.

Change-Id: If5e186d070ebc0e7c971ce786bc36d216339cd91