Update patch set 1
Patch Set 1:

Ian, thanks for creating the followup to add the additional log
files.

When you, Clark, and I discussed this in IRC, I did not realize
you were so opposed.  Please allow me to supply my reasoning.

I think it's great that people are enjoying the new journald
features in devstack.  As I said in IRC, my intent is not to
inhibit that in any way.

However, the production instance of nodepool we run for OpenStack
does not use the journal.  We have had some significant problems
with the journal on the few production hosts we use it
(unexplained complete absence of all log data, in short, is the
problem), so it may be a little while longer yet before we're
ready to start having our services log to it.  I personally find
it very useful to have some similarity between the log files I
see in our production instance and what I see in the nodepool
devstack runs.

Further, the log format output by the journal is not ideal.  I
appreciate that you are working to make it better, and once your
timestamp change lands and the no-hostname option shows up in a
distro we run, things may be much improved on this front.  But in
the mean time, given a choice between the two, I find the current
format more useful.

Finally, there is the aspect of the log content itself.  Much of
the particular debugging I do with nodepool relates to
interactions between multiple threads or processes.  Because in
our logging config the timestamp is added in the formatting stage
of the python logger, it provides some guarantees as to how it
relates to the sequence of commands within a certain thread of
execution.  Given an execution sequence of "A; log; B;" when the
timestamp originates in python, we can be certain that A happened
no later than the timestamp and B happend no earlier.  Knowing
this with some precision is often very useful when debugging a
threading problem.  In the case of journaled standard output in
devstack, the only guarantee we can make is that A happened no
later than the timestamp -- we can make no inferences about B in
that case because it is possible for the python process to resume
before the journal ingests the log entry.  In examining the
current output, it often appears that the journal timestamp is
delayed by about 1ms compared to the python timestamp, even under
no load, sometimes more under heavier load.

It is certainly still possible to debug thread interactions under
these conditions, but it is very important to be aware of what
exactly the log statements mean with respect to the program.  The
journal timestamps do mean something different than the python
timestamps, and at least for the moment, I find the python
timestamps to be useful in the kind of debugging I am called upon
to perform.

Please note that all of that is merely a defence of how I use the
log files.  I understand that the journal provides other features
that are useful in other kinds of debugging.  Again, I'm not
suggesting that we do this *instead* of the journal, just that we
*also* keep the facility we have now if it's not too onerous.

Patch-set: 1