Update patch set 5
Patch Set 5:

I haven't looked closely into either this or the suggestion from Shrews yet, but just wanted to leave a note so that folks know the purpose of this code is to move an exception from one thread (the provider manager) into another (the launcher).  The use case is not entirely straightforward, and it's worth careful examination of the old and new behavior.

Patch-set: 5
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0