Update patch set 2
Patch Set 2:

I don't mind if it's in or out of tree (building the image is possible with the same file in both ways). So I'm ok with leaving this out of tree. Just thought it might be interesting for others as I saw this topic several times on irc (and everyone crafting his own image).

Patch-set: 2