Don't log tracebacks for server-side missing image
Sometimes we request booting a server with an image that the cloud
cannot find. It is definitely an error that needs to be reported when
this happens. However, it's not an error that the traceback provides
additional information on.

We can provide a more useful error in this case by writing one directly
rather than by propogating the traceback.

Change-Id: I49f7e8d86d1fe0cc7180fc9390d8c2c33c95ad1a