Support booting cloud-images by name or id
The docs say we support this, but the code doesn't.

Also, self._cloud_image.name == self._label._cloud_image and is
essentially a foreign key. That's hard to read at the call site, so just
use self._cloud_image.

We have a cloud id if it's a disk image- so wrap that in a dict. Pass
the other one through unmodified so that we'll search for it.

We also don't have any codepaths using image_name, nor a reason to
distinguish.

Change-Id: I4aa9bd8e7c578ae63d05df453b9886c710a092c0