Update patch set 1
Patch Set 1: Code-Review-1

We should call this label -- I'm going to push up a change to the zk schema shortly to switch 'type' to label.  We started using 'type' when I thought we were going to get rid of the term 'label'.  But we ended up keeping it, so we should re-standardize on it.

Patch-set: 1
Label: Code-Review=-1