Restore separate upload logs
In I2ac1ec913e071f1f58830f8f2beebb88176b3dc7 we added a split-out of
upload logs which has been lost in subsequent updates.

The motivators for this are less in a nodepool-builder world given
less is happening on a host.  However, we still have all these loggers
setup in the configuration, and importantly, they're set to debug
level.  When diagnosing an upload issue it's much easier to be able to
examine just the specific log file, without everything else mixed in
between.

This just restores using the individually named logger in the upload
function.

Change-Id: If952c373fcb98afc99a006492949f9468c8a2460