Update patch set 7
Patch Set 7: Code-Review-1

(4 comments)

I think this is looking great. The only things I think are -1 worthy is the get_provider_manager comments. That said, I do think that if we can fully use the driver interface to get the fake providers in the test suite, it'll actually be a nice way to show the system working. (there are several hacks in place with the fake providers that are only there because there isn't a full-on interface, so now that you're adding one, let's just use it)

Patch-set: 7
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1