Update patch set 9
Patch Set 9: Code-Review-1

(1 comment)

Similar to the previous patch- I'm finding it hard to fully reason about this without looking at a config example.

Perhaps we should etherpad up the config so we can iterate on what we think it should look like quickly and easily.

I also think (and maybe we should put this into the spec) that the flow needs to be:

- nodepool-launcher starts up, does a leader-election check in
  zk
- if not leader - nothing is done with config
- if leader, reads node config from file and updates zk

Then, during operation, node check-out and check-in is always done on entries in zk.

The other option would be a specific utility "load static node config into zk" that could be triggered as part of config management, as well as "add node" and "remove node" commands, so that static launchers never actually read config. That would remove the need for leader election and also would allow safe updating of config from one place only (so that one doesn't need to worry about a launcher crashing, staying gone for a while, coming back with out of date config and then becoming the leader - which can be handled, but is a lot of work)

Patch-set: 9
Label: Code-Review=-1