Use threading.Event for quicker shutdowns
Right now it is possible for a shutdown event
to be reacted to in the order of tens of seconds
due to the usage of time.sleep.

Using an event allows this to be reduced to nearly
instanenous and instead of using time.sleep the internal
event.wait is used as the thing to sleep on instead (and
it will eject from sleeping the moment the event is set).

Change-Id: Iea6c7cc0902db7c7efdf740c3fbdc870bb85f0ae