Update patch set 1
Patch Set 1:

So removing secure altogether was the first patch i was going to do, but came across [1] which would seem to imply that we know it doesn't do anything but are saving it for some purpose. I'd be happy with removing it. 

[1] https://github.com/openstack-infra/nodepool/blob/feature/zuulv3/doc/source/installation.rst#configuration

Patch-set: 1
Reviewer: Gerrit User 7191 <7191@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0