Update patch set 1
Patch Set 1:

The plan was to use the secure file for zookeeper credentials at some point, so specifying the file to the daemons was made optional. I'm not entirely sure where we stand with that going forward, so this may need some deeper discussion before we totally remove things.

Patch-set: 1
Label: Verified=0