Update patch set 1
Patch Set 1:

>> Q: why could we not store this information into zookeeper? Or is it a chicken and egg issue?

Yeah. We'd have to use the hostname to identify the znode that contains the UUID for that builder. But if hostname changes, we've lost our znode with that info. So, same issue, really. Local storage avoids that problem.

>>> Follow up, would pulling in something like dogpile.cache.dbm also make more sense?

Does that give any benefit over a static file here?

Patch-set: 1
Label: Verified=0