Update patch set 18
Patch Set 18: Code-Review-1

We should still circle back with Monty and discuss the approach.  I think Monty and I were imagining the Node znodes would be long-lived, but this patch implements transient Node instances.  I'll need to think about the pros and cons of each.

Regardless, this should probably come with docs.  We could either revise this patch, or add them as a followup, but I don't think we should land it without them.

This does look like a good implementation of the transient-node approach.

Patch-set: 18
Label: Code-Review=-1