WIP: Fix broken use of pre-existing cloud images
The existing code to check for a label being ready was using the label
name to query the cloud for image existence. The label name and the
name used to communicate with the remote cloud are not necessarily the
same.

Change the config reader to put the ProviderCloudImage into the Label's
cloud_image attribute, which allows validating that the cloud-image
referenced in a label definition matches the name of an existing
ProviderCloudImage.

Add accessor properties on the ProviderCloudImage to get the thing we
pass to the cloud (external) and the human readable name version of the
name of the image on the remote cloud for logging (image_name).

Note: This will need to be reworked a bit with the upcoming pluggable
config. The name, image_id, image_name split is done the way it is
currently because of how shade works and so is likely a feature of an
OpenStackProviderCloudImage not a general CloudImage. (or maybe I'm
wrong and it's a general thing, but I don't think I'm wrong)

Change-Id: I8411c627f9136339d1b0eb35632d6b2a27ab7a81
Co-Authored-By: Monty Taylor <mordred@inaugust.com>
WIP: this needs a unit test added