Don't fail on quota exceeded
When running nodepool with labels of different size it is not possible
to adjust max_servers such that all resources can be used without
running into quota on the provider. In this case after the retry limit
the node allocation fails. This can be fixed by not treating 'quota
exceeded' as a permanent error and don't count this for the retry
limit.

Change-Id: I2fb2ea3e0d68cfa52b2b663b9c6d1654e2f45e97