Support several paused request handlers
In order to support quota failures more gracefully we need to pause
the request handlers when a quota error is discovered. These are
caused after the max-servers check which pauses the request handler if
this is exceeded. So this is laying some groundwork such that request
handlers also can be paused by node launchers getting quota errors
back.

Change-Id: I5410a9108672e7030750a6754b291a7366226c6a