Add cloud quota handling
This is an approach of calculating the cloud quotas before launching
nodes. To support co-existence with other services the limits of the
cloud are gathered and all unknown resources subtracted from the max
values. Thus the remaining quota for being used by nodepool can be
calculated and checked before launching nodes.

Change-Id: I18594ab76949faddade3164b6027dd3b82771c95