Provide resource_name in logging as structured data
Appending resource names to loggers is great for routing log messages on
a per-resource basis, but it means the fully qualified logger name has
to be included in a log line in order to make use of the resource name.

Add a LoggingAdapter that adds resource_name to the LoggingAdapter extra
dict so it is available in format strings.

Append the resource_name to the end of the logger name as is
currently done, and add a short_name to the extras dict that does not
have the resource_name appended so that log messages can be constructed
with the logger name and resource name called out separately.

Change-Id: I218ae4d84331e69dd0cc21d3f14c0feeecdf2dac