Do not satisfy min-ready requests if at capacity
If a provider gets a min-ready request, but it is already at capacity,
it should decline the request so as to not wedge the provider (because
a min-ready does not reuse READY nodes, provider would have to wait until
a node was freed to satisfy the min-ready request).

Also, let's make min-ready requests the highest priority so that the
requests are satisfied nearer to the time when it is discovered that
there are not enough ready nodes.

Change-Id: I37dc0802c1b30714833d6f46cc19b86deb58852b