Update patch set 1
Patch Set 1:

I wonder if this should be normal, or even below-normal priority.  Min-ready is designed to keep the system responsive by not having to wait for nodes for a request, but if a real request is already there and the nodes aren't, satisfying min-ready ahead of it isn't going to make anything run faster.

Patch-set: 1
Label: Verified=0