Update patch set 7
Patch Set 7: Code-Review-1

I think, if we go with this, we should default the username to 'zuul' so we don't break our existing production setup which does not have the username configured in its config files. That would, conveniently, also eliminate the need to change all of the test .yaml files.

Patch-set: 7
Label: Code-Review=-1