Run image object autocleanup after uploading images
shade should auto-cleanup objects created for image uploads, but we all
knows things go wrong. After uploading an image, run the object
autocleanup to ensure that we reap any objects that slipped through.

This is added to the upload worker and not the cleanup worker because we
need to make sure not to run the autocleanup while an image is being
uploaded, because the cleanup has no way of knowing if a given object is
in the middle of being used for an import in a different thread.

Change-Id: I021610506b782fa62e2fbb517beea5d7e7733d73
Depends-On: Ifb697944856e1922517074d84a7c00a4af75b1e6