zk: use kazoo retry facilities
This change wraps all the zookeeper calls with the retry helper so that
ConnectionLost and SessionExpiredError are handled transparently. This
makes Nodepool service resilient to zookeeper connections issues.

This change also removes the connection listener and uses a new checkLock
method to ensure Lock is still valid. This fixes a missing check in the
launchNode code.

Nodepool now waits indefintely for Zookeeper service to become available,
and it logs retry attempts every 10 seconds.

Change-Id: Ifaf1fb9f6879c57faaaf490a705ea78afef44420