Add a plugin interface for drivers
This change adds a plugin interface so that driver can be loaded dynamically.
Instead of importing each driver in the launcher, provider_manager and config,
the Drivers class discovers and loads driver from the driver directory.

This change adds a drivers-dir option to the nodepool.yaml so that driver
can be loaded out of the tree.

Change-Id: Ia347aa2501de0e05b2a7dd014c4daf1b0a4e0fb5