Handle race between handler and request cleanup
Our node request handlers are polled for completion. There is the
possibility that the request belonging to a request_handler has already
been cleaned up by the cleanup worker because its work has been
completed. If the request handler is polled for completion after the
request and its corresponding request lock have been cleaned up we will
throw an error and fail to poll other request handlers for completion.
This leads to a request_handler leak (though requests should all
continue to be cleaned up properly by the cleanup worker).

We can address this by specifically checking for the case where a
request lock does not have a znode when polling the request handler for
completion. If there is no lock then there isn't anything more to
cleanup in this case so we log it and move on.

Change-Id: I30e78bc67906d9ad97bf9ca964ea145c67526a8b