Update patch set 7
Patch Set 7:

Just started to look into that. Externally deleting the node request and waiting for the lock deletion does not reset the lock property of the cached noderequest. Thus it releases a non-existing lock which is handled gracefully in kazoo. So this does not seem exactly to be the way to reproduce this.


The actual error is thrown when nodepool itself unlocked this somewhere else (which involves setting lock to none). But I haven't yet wrapped my head around it.

Patch-set: 7