Update patch set 7
Patch Set 7: Code-Review-1

After looking more into this test, I think the larger issue (that this does fix, but does at least handle) is that the request.lock kazoo Lock either gets reset to None, or accidentally overwritten (maybe because of object references?).

Since I'm afraid of us getting into the same situation we saw in production because of this (lots of READY and locked nodes not getting handled by the request handler b/c of the exception being thrown), I'm going to propose we proceed with this, but:

  1) remove this test since it isn't valid

  2) remove the newly added exception (and just catch ZKLockException inside poll() method) since the new exception isn't really getting us anything and maybe adds some confusion.

With this in place, it will eliminate the deadlocked behavior that caused us issues and gives us time to hunt down the real problem with the missing lock.

Comments?

Patch-set: 7
Label: Code-Review=-1