Clarify provider manager vs provider config
There are places where we handle both the provider config and provider
manager objects. We had a bug where we called a manager method on the
config object which failed. Since python doesn't type check ahead of
time let us try and make this easier on developers and very explicitly
name the provider manager provider_manager and the provider config
provider_config.

Change-Id: I36919672cdbf9a5c66297ff49f304b9040eee1cb