Add DIB faq to docs
Periodically we get asked to just use someone elses
$favoriteimagebuildtool. There are good reasons for using
diskimage-builder, but they may not be apparent to everyone so add a
small doc explaining those reasons.

Hopefully this makes people more comfortable with the tool choices and
at the very least gives us something to point to that we can evaluate
other tools against when asked why we aren't using them.

Change-Id: I8c8311469fc6fe0b2d3661143001d15abe8aee59