Only fail requests if no cloud can service them
Previously if we had cloud errors when attempting to handle a request in
a particular cloud we treated that node request as failed even if other
clouds had still not attempted to fulfill the request. This is
problematic because clouds have outages and this is the reason nodepool
supports speaking to more than one cloud at a time.

We fix this by only marking the request failed if all other providers
have attempted to fulfill it. Otherwise we put it back in the requested
state and let other providers have a go.

Change-Id: I519cdd2401f77cad97d60329e207dddafc8cd5a4