Add test_launcher test
The aim of this test is to test that one provider handler being in
failure mode due to something liek a cloud disappearing doesn't fail
node requests as long as other providers are able to service the
request.

Problem is each provider handler runs in its own thread and this is
incredibly race prone. Haven't quite sorted out how to handle that yet.

Change-Id: Ia009ac0d16d311d3006f5910ad981900a264a9a8