Update patch set 36
Abandoned

The 'feature/zuulv3' branch has merged into 'master'.  In order
to remove the old branch, we must close all of the open changes.
This technicality is the only reason this change is being
abandoned.  Please re-propose this change to the master branch
and it will be considered.  When you do so, please include a link
to the original change in Gerrit to assist reviewers.

Thank you for your patience and assistance.

Patch-set: 36
Status: abandoned