Update patch set 1
Patch Set 1:

As described in the link from my original comment (https://cwiki.apache.org/confluence/display/ZOOKEEPER/ErrorHandling), a non-idempotent request is a request that could end up with a different result when re-issued. For example, we depend on sequences quite a bit. With this retry, it seems possible we could re-issue a create() of a sequence node and end up with two znodes. I'm not certain that we'd handle that very well.

Patch-set: 1
Label: Verified=0