Switch to stestr
stestr is maintained, while testrepository is not. OpenStack is
migrating to stestr, go ahead and follow suit.

The --no-subunit-trace flag is added to keep the quiet test run
behavior. If we remove it, stestr will by default print the name of each
test as it starts it.

This uses environment inclusions which need tox v3, so increase the tox
minversion.

Change-Id: I08cd4bee86b56d9a68644b049497c34e34e5f226