Fix relaunch attempts when hitting quota errors
Currently when hitting quota errors we just invalidate the quota cache
and immediately try again to launch the node. Of course that will fail
again in most cases. Instead we need to delete the according zNode
such that the handler which has quota and pause handling just
reschedules a new node launch once the quota allows for it.

Change-Id: I122268dc7723901c04a58fa6f5c1688a3fdab227