Fix relaunch attempts when hitting quota errors
Currently when hitting quota errors we just invalidate the quota cache
and immediately try again to launch the node. Of course that will fail
again in most cases. Instead we mark the node as aborted which is no
failure indicator and pause the handler it will automatically
reschedule a new node launch once the quota allows it.

Change-Id: I122268dc7723901c04a58fa6f5c1688a3fdab227