Update patch set 1
Patch Set 1: Code-Review-1

Hrm, if I read this correctly, I don't think your first paragraph in the commit message is correct here.

Because you use deleteNode(), you're actually bypassing our API and directly deleting the ZK node (not the instance). The cleanup thread will see this as a leaked instance and delete the instance, but there is no DELETING state recorded. FWIW, if you don't want to bypass the API to delete it, proper thing to do is just set state=DELETING and let the cleanup thread handle it.

Patch-set: 1