Update patch set 5
Abandoned

Actually due to the change to the v3-native devstack job, we may want
to abandon this... so...

Hi,

We recently merged the feature/zuulv3 branch back into master.
This includes many significant changes to the behavior and
underlying implementation in Nodepool.  It appears this change has
either already been implemented in that branch, or is no longer
applicable.

Thank you very much for the contribution, and if this analysis is
in error, please accept my apologies and feel free to restore
this change and rebase it on the current tree.

Patch-set: 5
Status: abandoned