Update patch set 8
Patch Set 8:

> This should be proposed under the zuul CLI command. We should have
 > removed the 'hold' command for the nodepool CLI long ago. It's
 > basically worthless in this version of nodepool. Zuul controls when
 > a node is held.

I've added this on the Zuul side: https://review.openstack.org/#/c/539596/ which adds an option to the autohold command to specify for how long a node set should be on hold before being automatically deleted.
Nevertheless I do believe this patch has a place within nodepool for the following reasons:

- there's a similar feature for "ready" nodes
- a node on hold is an "abnormal" state, meant for debugging, and it is costly in resources. This automated cleanup simplifies the lives of operators with the guarantee that held nodes will be deleted at some point.
- the patch expands a pre-existing cleanup routine in nodepool which I haven't seen in Zuul (please correct me if wrong), so why reimplement it in Zuul?

As for nodepool's CLI, it is addressed in a follow-up patch to allow overriding the configuration setting: https://review.openstack.org/#/c/539493/

I can remove the CLI part if necessary.

Patch-set: 8
Label: Verified=0