Update patch set 8
Patch Set 8: -Code-Review

Sorry, this comment was supposed to be on the next review in the chain (https://review.openstack.org/539493). Nodepool should still be the one to clean up held nodes, not zuul. We just want folks to be able to set a hold via the nodepool command.

Patch-set: 8
Label: Code-Review=0