Update patch set 3
Patch Set 3: Code-Review+2

There's no reason we can't pick these up in the devstack job, however, the devstack job is currently a work in progress that other folks are working on, and it appears they haven't gotten picking up *any* logs working at the moment, so I can't really address that.  I'd like to avoid blocking on that.

Patch-set: 3