Preserve logging handlers across dameonize
If you have a mistake in your logging config (unfortunately easy to
do) you currently get very little help.  "service nodepool-builder
start" gets a valid return (as the process has daemonized), but since
the logging is not set up, any exception from the logging
configuration is eaten by daemon redirects.  The result is that the
process reports being started OK, but dies instantly and silently.

This is a fair pain; if you're lucky you might start the builder
interactively and see it, but otherwise you end up having to "strace
-f -p 1" and start pulling apart the systemd launching of the process.

A semi-common trope seems to be to find the list of file descriptors
being used by the logger, and ensure these are preserved across
daemonization (e.g. [1], [2]).  By doing the setup before
daemonisation, the service/init.d start should fail helpfully if there
is a logging config error.

[1] https://github.com/openstack/freezer/blob/master/freezer/scheduler/daemon.py
[2] https://stackoverflow.com/a/24298373

Change-Id: Ib0b67485778e99316d1f82bc59a47b3643760e45