Consolidate status results
All the _list functions return the same thing; save the results and
use a common output function to generate the json or text output.

label_list was the only function that didn't follow the pattern of
returning a list of objects.  I'm not sure that for JSON output it
makes any sense to return repeated dicts just with label names; turn
this into a list of labels and their current count which seems
possibly useful.

Change-Id: I9cb44b09de2cb948e7381ef10302b21040433a2c