Store label info with launcher registration
It can be useful to know all of the labels that nodepool can support
without having to examing nodepool configuration files. We can store
this information when we register each launcher at startup, and on
configuration changes.

This stores supported labels for each launcher in ZooKeeper. It also
introduces a new model class (Launcher) to extract away the information
that we store in ZooKeeper so that we can later more easily add other
information that could be useful (such as per-tenant info).

Change-Id: Icfb73fbe3b67321235a78ea7ed9bf4319567eb1a