Update patch set 1
Patch Set 1: Code-Review-1

A general concern with this as currently implemented is that it doesn't appear that we will only register labels that can be booted by the launcher. I think this can happen if none of the providers are capable of booting the label or if a provider is later updated to remove label support.

Right now this is "ok" because we don't hvae a strong contract with zuul about available bootable labels but this change changes that to make it much more verbose. I think we should do our best to only register the labels that can be booted if we are making this a proper api.

Patch-set: 1
Label: Code-Review=-1