Add nodepool service file for CentOS7
The existing service file is for Fedora27 and it does not work
on CentOS7 since pip installs nodepool in /usr/bin on this distro.
This change adds the service file for CentOS7 and moves the
existing one under its own folder named after the distro.

This change is uploaded to collect feedback for zuul-from-scratch
CentOS7 installation instructions which will be uploaded soon.

Depending on the feedback, this change could either be abandoned or
amended to reflect the received feedback to the change itself and the
corresponding documentation.

Change-Id: Id6a6303d4785b64d9d259b9d553dea6d07eeeedb