fail to build image without known type
It is described in the docs, that nodepool will build image only if it
is referred in one of the clouds, but it is possible to force that
creation: `nodepool image-build fedora-28`. This leads to inability to
define image types and empty argument to DIB (causes late failure): `-x
-t --checksum --no-tmpfs -o /opt/nodepool/images/fedora-28-0000000001
fedora-minimal vm simple-init nodepool-base cache-devstack
initialize-urandom growroot infra-package-needs`. Raise exception early
if the type is not known.

Change-Id: I27f8f45ba62e5ee0cca0ea7cd27893092b726ebd
Story: 2001963