Update patch set 1
Patch Set 1: Code-Review-1

Thanks, I agree you've found a problem here.  However I think failing at this point is probably not the most helpful way and somewhat opposite to the idea that we automatically figure out the formats as it mentions in the docs.

my first thought is maybe just

 img_types = ",".join(diskimage.image_types) if diskimage.image_types else 'qcow2'

which would replicate the dib default, which is .qcow2 (or, in config.py do the same thing essentially and set the default there if nothing else sets it)

Patch-set: 1
Label: Code-Review=-1