Fix ConfigValue comparisons
Our code was terribly inconsistent with implementing the __eq__()
method for ConfigValue based objects (if we did at all). If this
was implemented, not all of the attributes were used in comparison.

This change forces a ConfigValue derived object to implement the
comparison method, and fixes the places we neglected attributes.

Adds some very simple tests to at least exercise this code.

Change-Id: Ia2f600a9a4f3770087372bbc9f07531d5ea569e1