WIP: Simplify driver API
Further decouple the driver API from the provider configuration
by simplifying the driver interface (which will eventually lead to
the driver deciding whether or not to support multiple labels).

This significantly simplifies the driver API by reducing the launching
of nodes to a launch() and poll() interface. How the drivers actually
launch nodes is decoupled from the driver API (leaving it entirely up
to the driver to decide the best way to do that). The poll() method
acts as an asynchronous check on whether the node set launches have
completed in their entirety.

The NodeLauncher class (currently used only by the OpenStack driver),
is moved to a driver utility module that other drivers may choose to
use or ignore.

Change-Id: I407d999b3608e1614c0dbe49808b2a64756dde58